Fix/p2p grpc error (#368)
Fixes #281 

Its too annoying to restart the builds and develop seems to fail almost everytime due to this. I pushed a fix to upstream. Until the [PR](https://github.com/paralin/go-libp2p-grpc/pull/1) is accepted there using my fork as the source.

Once upstream accepts the PR, we can use that in the dep